refactor: move MQTT logic to dedicated module
